Composer+PHPUnit standards
