+++ composer.json
@@ -1,0 +1,23 @@
+{
+    "name": "shuber/curl",
+    "description": "A basic CURL wrapper for PHP.",
+    "type": "library",
+    "keywords": ["php", "curl"],
+    "authors": [
+        {
+            "name": "Sean Huber",
+            "email": "shuber@huberry.com"
+        }
+    ],
+    "require": {
+        "php": ">=5.3.2"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "3.7.*"
+    },
+    "autoload": {
+        "psr-0": { "shuber\\Curl": "src" }
+    },
+    "minimum-stability" : "stable"
+}
+
