--- curl.php
@@ -1,4 +1,0 @@
-<?php
-
-require_once 'lib'.DIRECTORY_SEPARATOR.'curl.php';
-require_once 'lib'.DIRECTORY_SEPARATOR.'curl_response.php';
