+++ tests/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+
+$loader = require dirname(__FILE__) . '/../vendor/autoload.php';
+$loader->add('Curl\Test', __DIR__);
+$loader->register();
